Con esto esta listo EditarDB -> Materias -> Details

+++ app/src/main/java/com/wolfteam20/schedulemobile/data/network/models/MateriaDTO.kt
@@ -4,11 +4,11 @@ package com.wolfteam20.schedulemobile.data.network.models
  * Created by Efrain Bastidas on 2/3/2018.
  */
 data class MateriaDTO(
-    var codigo: Int,
+    var codigo: Long,
     var asignatura: String,
     var horasAcademicasTotales: Int,
     var horasAcademicasSemanales: Int,
-    var idSemestre: Int,
-    var idTipo: Int,
-    var idCarrera: Int
+    var idSemestre: Long,
+    var idTipo: Long,
+    var idCarrera: Long
 )
